requirements.txt
streamlit
boto3
opencv-python-headless
Pillow
pandas
